Reorganization of folders medWidgets PART 2
