+++ src/layers/medCore/source/settings/medSourcesLoader.cpp
