+++ src/layers/medCore/source/settings/medSourcesLoader.h
