+++ src/layers/medWidgets/source/settings/medSourcesLoaderPresenter.cpp
