+++ src/layers/medWidgets/source/settings/widgets/medSourcesSettings.cpp
